Update patch set 7
Patch Set 7:

This new repo does not cause any new bugs to appear and it fixes a lot, this is also the proper format for this repo because 5422 should never have been converted to 5420. The kernel is a 3.10 we have a 3.4 the mobicore driver is a newer version that our kernel does not support.

Patch-set: 7