Update patch set 7
Patch Set 7:

We shouldn't be forced to use a incompatible platform repo the mobicore is newer and we have no way of moving to a 3.10 kernel. The gpu crash is related to gralloc. This new repo is using the same format as what 5410 did because like it the new code doesn't support 5420 out of the box and 5422 is just different enough to matter

Patch-set: 7