Update patch set 7
Patch Set 7:

Wait, I'm confused, so you couldn't get it to include properly, so you never managed to test with the exported kernel header, or you did manage to export it and test with the exported kernel header (which then did not work)?

Either way, wholesale replacement of this code is wrong.

Patch-set: 7