Update patch set 7
Patch Set 7:

We have tested with the proper header being included through the repo by just replacing it here it did not fix the problems. I think wholesale replacing it is better because this version is too new and we don't support it so we have to downgrade it. This newer version was not designed for 5420.

Patch-set: 7