Update patch set 7
Patch Set 7:

I mainly compared the mobicore kernel drivers of the 5420 and the 5422 this shows that we do not implement the secure storage or the FSD (File system driver). This confirms what the logs said when they complained about the two of those failing to start. It is also very hard to get detailed logs for mobicore #blamesamsung

Patch-set: 7