Update patch set 7
Patch Set 7:

I am not going to waste my time trying to support a newer version of mobicore that is designed for a newer soc and a newer linux kernel version. Just so we can use the improper soc hardware repo. Thanks but no thanks, if that is what you say i need to do i will just drop support for the klimt's and chagall's officially and make unofficial's where we can use the the proper hardware repos for our devices.

Patch-set: 7