Update patch set 7
Patch Set 7:

It can be pretty well summed up by saying this version does not match what's in the kernel so there is going to be mismatches all over the place struts will change. The logging mechanism for mobicore blows most errors are not printed so i at least can not track down exactly what is breaking in the code but the fingerprint scanner does not function with the new mobicore where as with the original code designed for 5420 it works fine.

Patch-set: 7