Update patch set 7
Patch Set 7:

> If you could tell me, "yes, it's because this struct in this header
 > has changed", or "yes, it's because it's calling a function in the
 > kernel that now has a mismatched parameter", then I can help you
 > out.

The things is it is widespread.  Across 6 different tablets for sure and I am guessing the Exynos Note 3 is also affected.  The mobicore changes breaks mc_Linux.h.  It doesn't matter if it is called in kernel or hardware it just does not work or build for that matter.  Why we switch to the new exynos5422 repos I do not know because the insignal hardware worked just fine.  All I know is someone did a half ass job to begin with now we have to deal with it.  Seriously, three of us are saying that there is detectable changes in performance and other issues.  

Why would switching back to an older version of hardware be very disparatible?  Sean and I have either talked to other exynos5420 maintainers or are the maintainers for 90% of the devices affected.

Again why in the world we were forced to use this misconfigured repo to begin with is behind me.  I said a year ago that this was a dumbass idea now fast forward a year and here are as a group of device maintainers saying it does not work.

Patch-set: 7