Update patch set 7
Patch Set 7:

As you can see from above the current repo does not have a functioning mobicore you also get to see how little logging there is. But i know what your gonna say that's an easy fix just add that missing .tlbin, and i will say that .tlbin file does not exist for the 5420 on any of the stock firmwares i have looked for it. He is running in permissive so those denials would not affect it either.

Patch-set: 7